Support for new Gaudi cmake
- Delegate as much as possible to the LbDevTools toolchain.
- No support yet for distcc when building Gaudi.
- No support yet for limiting the number of parallel genreflex jobs.

+++ project.mk
@@ -152,7 +152,7 @@ ifeq ($(VERBOSE),)
 # (emulate the default CMake install target)
 install: patch-python-ns
 	cd $(BUILDDIR) && $(CMAKE) -P cmake_install.cmake | grep -v "^-- Up-to-date:"
-	cp $(BUILDDIR)/config/$(PROJECT)-build.xenv $(INSTALLDIR)/$(PROJECT).xenv
+	test -f $(BUILDDIR)/config/$(PROJECT)-build.xenv && cp $(BUILDDIR)/config/$(PROJECT)-build.xenv $(INSTALLDIR)/$(PROJECT).xenv || true
 endif
 
 # ensure that the target are always passed to the CMake Makefile
